Floorplans have been researched into
Floorplans for horror games have been researched into and recorded in this pressentation. However, it would seem most horror games don't have a hiding mechanic, atleast not as many as I though so we may need to change how we get this game to work. I will Check with a lecturuer if it still fits the brief as most succesful horror games include hiding, it's just the ones less known to the world; don't.